refactor: Simplify sidebar prop and remove Logout modal integration
Previously, two props were in place for authentication methods and modal visibility. In the current version, the focus shifts to a single prop named 'handleShowLogout,' responsible for triggering the display of the Logout modal in the 'LOGGED-USER' layout component when clicked in the logout icon.

Furthermore, this change involves removing the direct integration of the Logout modal from the sidebar. By simplifying the props and shifting the responsibility to the 'LOGGED-USER' layout, the codebase becomes more organized and aligned with best practices.

Author: Wellesley Mussolini