feat: Post interactions with hover effect
Refactored the 'Post' component to componentize post interactions, including like, comment, share, and send functionalities. This change not only enhances the code but also adds a hover transition effect to the post interactions, improving user interactivity and visual feedback. It will make it easier to manage and customize post interactions.

Author: Wellesley Mussolini