feat: Introduce Post Button Component
Created a new 'PostButton' component to facilitate the publication of text commentaries in the database. The component interacts with the server file to execute backend functions using Prisma. It includes a feature to block posting if the user hasn't entered any text, and when a post is made, the button transforms into a loading spinner to provide visual return while the process is being published.

Author: Wellesley Mussolini